This is the initial upload of Truman New Template
This code was modeled after teh original Truman.

It was designed to be run via Orchestrator in unattended mode.

The config file and process files are stored here as example and will need to be updated with the correct information when run. The code is using a shared drive. The share drive will contain the config and process file.